+++ include/mamba/core/context.hpp
@@ -189,6 +189,7 @@ namespace mamba
         bool override_channels_enabled = true;
 
         std::vector<std::string> pinned_packages = {};
+        bool freeze_installed = false;
 
         bool use_only_tar_bz2 = false;