+++ src/micromamba/install.cpp
@@ -1,5 +1,6 @@
 #include "common_options.hpp"
 
+#include "mamba/api/configuration.hpp"
 #include "mamba/api/install.hpp"
 
 
@@ -10,6 +11,12 @@ void
 set_install_command(CLI::App* subcom)
 {
     init_install_options(subcom);
+
+    auto& config = Configuration::instance();
+
+    auto& freeze_installed = config.at("freeze_installed").get_wrapped<bool>();
+    subcom->add_flag(
+        "--freeze-installed", freeze_installed.set_cli_config(0), freeze_installed.description());
 
     subcom->callback([&]() { install(); });
 }